creating database functions
